#8 rename from ASID to ASIP///spell checking///modulorization
