+++ Dockerfile
@@ -4,8 +4,8 @@ WORKDIR /usr/src/app/
 
 COPY requirements.txt ./
 RUN pip install --no-cache-dir -r requirements.txt
-COPY asid_v2 asid_v2
-WORKDIR /usr/src/app/asid_v2
+COPY asip_v2 asip_v2
+WORKDIR /usr/src/app/asip_v2
 
 
 ENTRYPOINT [ "python" ]
